requirements.txt
paho-mqtt==1.2
requests
